refactor: RandomGenerator를 common.js로 이동
RandomGenerator가 따로 파일로 분리되기에 너무 작다.
따라서 common.js 파일에서 유틸함수의 느낌으로 사용하는 것이 더 좋을 것이라고 판단한다.
