--- src/model/RandomGenerator.js
@@ -1,7 +1,0 @@
-const RandomGenerator = {
-  getBetween(min, max) {
-    return min + Math.floor(Math.random() * max);
-  },
-};
-
-module.exports = RandomGenerator;
